# src/main/resources/static/openapi/onboarding-api.yaml
openapi: 3.0.3
info:
  title: Onboarding Orchestrator API
  version: 1.0.0
paths:
  /api/v1/onboarding-cases:
    post:
      operationId: startOnboarding
      summary: Start an onboarding case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartCaseRequest'
      responses:
        '201':
          description: Case created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/onboarding-cases/{caseId}:
    get:
      operationId: getOnboarding
      summary: Get an onboarding case
      parameters:
        - $ref: '#/components/parameters/CaseId'
      responses:
        '200':
          description: Current case state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/onboarding-cases/{caseId}/interaction:
    get:
      operationId: getCurrentInteraction
      summary: Get the current interaction
      parameters:
        - $ref: '#/components/parameters/CaseId'
      responses:
        '200':
          description: Current workflow interaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/onboarding-cases/{caseId}/actions/{action}:
    post:
      operationId: executeInteractionAction
      summary: Execute the action allowed by the current interaction
      parameters:
        - $ref: '#/components/parameters/CaseId'
        - name: action
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/ActorId'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Updated case
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Action is not valid for the current step
  /api/v1/onboarding-cases/{caseId}/events:
    get:
      operationId: listCaseEvents
      summary: List the immutable transition history of a case
      parameters:
        - $ref: '#/components/parameters/CaseId'
      responses:
        '200':
          description: Ordered case events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CaseEvent'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/onboarding-cases/{caseId}/consents:
    get:
      operationId: listCaseConsents
      summary: List the consent evidence registered for a case
      parameters:
        - $ref: '#/components/parameters/CaseId'
      responses:
        '200':
          description: Case consents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Consent'
  /api/v1/onboarding-cases/{caseId}/file-sets:
    get:
      operationId: listFileSets
      parameters:
        - $ref: '#/components/parameters/CaseId'
      responses:
        '200':
          description: Case file sets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileSet'
    post:
      operationId: createFileSet
      parameters:
        - $ref: '#/components/parameters/CaseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFileSetRequest'
      responses:
        '201':
          description: File set created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileSet'
  /api/v1/onboarding-cases/{caseId}/file-sets/{fileSetId}/documents:
    get:
      operationId: listDocuments
      parameters:
        - $ref: '#/components/parameters/CaseId'
        - $ref: '#/components/parameters/FileSetId'
      responses:
        '200':
          description: File metadata
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
    post:
      operationId: registerDocument
      summary: Register metadata for a binary stored outside the database
      parameters:
        - $ref: '#/components/parameters/CaseId'
        - $ref: '#/components/parameters/FileSetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDocumentRequest'
      responses:
        '201':
          description: Document metadata registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
components:
  parameters:
    CaseId:
      name: caseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    FileSetId:
      name: fileSetId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 150
    ActorId:
      name: X-Actor-Id
      in: header
      required: false
      schema:
        type: string
        maxLength: 150
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        maxLength: 100
  schemas:
    StartCaseRequest:
      type: object
      required: [workflowKey]
      properties:
        workflowKey:
          type: string
          example: onboarding
    CaseResponse:
      type: object
      required: [id, workflowKey, currentStep, status, data, version, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        workflowKey:
          type: string
        currentStep:
          type: string
        status:
          type: string
          enum: [IN_PROGRESS, COMPLETED, DECLINED]
        data:
          type: object
          additionalProperties: true
        version:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    InteractionResponse:
      type: object
      required: [caseId, step, terminal, status, metadata]
      properties:
        caseId:
          type: string
          format: uuid
        step:
          type: string
        action:
          type: string
          nullable: true
          enum:
            - select-products
            - accept-terms
            - submit-personal-data
            - submit-kyc
            - submit-address
            - verify-contact
            - verify-identity
            - submit-enrollment
            - accept-contract
        terminal:
          type: boolean
        status:
          type: string
        metadata:
          type: object
          additionalProperties:
            type: string
    CaseEvent:
      type: object
      required: [id, caseId, sequence, eventType, resultingStep, metadata, occurredAt]
      properties:
        id: {type: string, format: uuid}
        caseId: {type: string, format: uuid}
        sequence: {type: integer, format: int64}
        eventType: {type: string}
        action: {type: string, nullable: true}
        previousStep: {type: string, nullable: true}
        resultingStep: {type: string}
        outcome: {type: string, nullable: true}
        actorId: {type: string, nullable: true}
        correlationId: {type: string, nullable: true}
        metadata: {type: string}
        occurredAt: {type: string, format: date-time}
    Consent:
      type: object
      properties:
        id: {type: string, format: uuid}
        caseId: {type: string, format: uuid}
        type: {type: string}
        documentVersion: {type: string}
        accepted: {type: boolean}
        actorId: {type: string, nullable: true}
        correlationId: {type: string, nullable: true}
        evidence: {type: string}
        acceptedAt: {type: string, format: date-time}
    CreateFileSetRequest:
      type: object
      required: [name, maxFiles, allowedMediaTypes]
      properties:
        name: {type: string, maxLength: 120}
        maxFiles: {type: integer, minimum: 1}
        allowedMediaTypes:
          type: array
          minItems: 1
          maxItems: 3
          items: {type: string}
    FileSet:
      type: object
      properties:
        id: {type: string, format: uuid}
        caseId: {type: string, format: uuid}
        name: {type: string}
        maxFiles: {type: integer}
        allowedMediaTypes: {type: string}
        createdAt: {type: string, format: date-time}
    RegisterDocumentRequest:
      type: object
      required: [objectKey, originalFileName, mimeType, sizeBytes, checksumSha256]
      properties:
        objectKey: {type: string, maxLength: 500}
        originalFileName: {type: string, maxLength: 255}
        mimeType: {type: string, maxLength: 150}
        sizeBytes: {type: integer, format: int64, minimum: 1}
        checksumSha256: {type: string, pattern: '^[a-fA-F0-9]{64}$'}
    Document:
      type: object
      properties:
        id: {type: string, format: uuid}
        fileSetId: {type: string, format: uuid}
        objectKey: {type: string}
        originalFileName: {type: string}
        mimeType: {type: string}
        sizeBytes: {type: integer, format: int64}
        checksumSha256: {type: string}
        status: {type: string, enum: [AVAILABLE]}
        createdAt: {type: string, format: date-time}
  responses:
    BadRequest:
      description: Invalid request
    NotFound:
      description: Case not found